Implement A* algorithm
Use the A* search algorithm to find the path the shortest travel time
between the head of the snake and the fruit.
The snake takes the first step on the path, then recalculates the
optimal path after each step.

Uses the Manhattan distance as a simple heuristic to estimate the
remaining distance on the path.

The algorithm does NOT take into account where it end up leaving the
snake's body, it only considers the shortest path to the current fruit.